test feature based file structure with eslint rule
