+++ src/App/frontend/nextsrc/features/Instantiation/components/CreateInstanceButton/createInstanceButton.module.css
@@ -1,0 +1,3 @@
+.startNewButtonContainer {
+  margin-top: 12px;
+}
