+++ src/App/frontend/nextsrc/features/Instantiation/components/instanceSelection/instanceSelection.module.css
@@ -48,10 +48,6 @@
   background-color: unset;
 }
 
-.startNewButtonContainer {
-  margin-top: 12px;
-}
-
 .descriptionParagraph {
   margin-top: 12px;
 }
