+++ src/App/frontend/nextsrc/features/Instantiation/index.ts
@@ -1,0 +1,17 @@
+import { InstanceList } from 'nextsrc/features/Instantiation/components/instanceSelection/instanceSelection';
+import { PartySelection } from 'nextsrc/features/Instantiation/components/partySelection/partySelection';
+import {
+  activeInstancesQuery,
+  instantiationQueries,
+  partiesAllowedToInstantiateQuery,
+  useActiveInstances,
+} from 'nextsrc/features/Instantiation/instantiation.queries';
+
+export {
+  instantiationQueries,
+  partiesAllowedToInstantiateQuery,
+  PartySelection,
+  activeInstancesQuery,
+  useActiveInstances,
+  InstanceList,
+};
