+++ src/App/frontend/nextsrc/routes/party-selection/party-selection.loader.ts
@@ -1,7 +1,17 @@
-import { partiesAllowedToInstantiateQuery } from 'nextsrc/core/queries/parties/parties.queries';
+import { redirect } from 'react-router-dom';
+import type { LoaderFunctionArgs } from 'react-router-dom';
+
+import { GlobalData } from 'nextsrc/core/globalData';
+import { partiesAllowedToInstantiateQuery } from 'nextsrc/features/Instantiation/instantiation.queries';
 import type { QueryClient } from '@tanstack/react-query';
 
-export const partySelectionLoader = (queryClient: QueryClient) => () => {
-  queryClient.ensureQueryData(partiesAllowedToInstantiateQuery());
-  return null;
+export const partySelectionLoader = (queryClient: QueryClient) => (_: LoaderFunctionArgs) => {
+  const userPartyId = GlobalData.userProfile?.partyId.toString();
+  if (!userPartyId) {
+    throw redirect('/');
+  }
+
+  queryClient.ensureQueryData(partiesAllowedToInstantiateQuery(userPartyId));
+
+  return userPartyId;
 };
